Rename User from UserView and UserComponent to DashboardView and DashboardComponent. Add simple screen-change logic
