+++ budget-binder-multiplatform-app/src/jsMain/kotlin/de/hsfl/budgetBinder/compose/login/LoginComponent.kt
@@ -24,7 +24,7 @@ fun LoginComponent(screenState: MutableState<Screen>) = withDI(di) {
         onLoginButtonPressed = { email, password ->
             viewModel.auth(email, password)
         },
-        onLoginSuccess = { screenState.value = Screen.User },
+        onLoginSuccess = { screenState.value = Screen.Dashboard },
         onChangeToRegister = { screenState.value = Screen.Register }
     )
 }
