+++ budget-binder-multiplatform-app/src/jvmMain/kotlin/de/hsfl/budgetBinder/compose/Router.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
 import de.hsfl.budgetBinder.compose.login.LoginComponent
 import de.hsfl.budgetBinder.compose.register.RegisterComponent
-import de.hsfl.budgetBinder.compose.user.UserComponent
+import de.hsfl.budgetBinder.compose.dashboard.UserComponent
 import de.hsfl.budgetBinder.presentation.Screen
 
 @Composable
@@ -13,6 +13,6 @@ fun Router(screenState: MutableState<Screen>) {
         is Screen.Welcome -> {}
         is Screen.Register -> RegisterComponent(screenState = screenState)
         is Screen.Login -> LoginComponent(screenState = screenState)
-        is Screen.User -> UserComponent(screenState = screenState)
+        is Screen.Dashboard -> UserComponent(screenState = screenState)
     }
 }
