+++ budget-binder-multiplatform-app/src/jvmMain/kotlin/de/hsfl/budgetBinder/compose/dashboard/UserComponent.kt
@@ -1,4 +1,4 @@
-package de.hsfl.budgetBinder.compose.user
+package de.hsfl.budgetBinder.compose.dashboard
 
 import androidx.compose.runtime.*
 import de.hsfl.budgetBinder.presentation.Screen
